+++ src/lib.rs
@@ -40,6 +40,7 @@ pub mod input;
 pub mod math;
 pub mod physics;
 pub mod platform;
+pub mod profiler;
 pub mod renderer;
 pub mod time;
 pub mod window;